+++ code/src/java/pcgen/core/RollingMethods.java
@@ -315,8 +315,8 @@ public final class RollingMethods
 				{
 					throw new ParseException("Values greater than " + Integer.MAX_VALUE + " not allowed.");
 				}
-				int iRolls = (int) Math.round(((Double) numberOfRolls).doubleValue());
-				int iFaces = (int) Math.round(((Double) faces).doubleValue());
+				int iRolls = (int) Math.round((Double) numberOfRolls);
+				int iFaces = (int) Math.round((Double) faces);
 				if (numToKeep == 0)
 				{
 					numToKeep = iRolls;
