+++ code/src/java/plugin/exporttokens/EqContainerToken.java
@@ -195,7 +195,7 @@ public class EqContainerToken extends Token
 			}
 			else if (property.equals("RANGE"))
 			{
-				retString = Integer.toString(EqToken.getRange(pc, eq).intValue());
+				retString = Integer.toString(EqToken.getRange(pc, eq));
 			}
 			else if (property.equals("SIZE"))
 			{
@@ -229,7 +229,7 @@ public class EqContainerToken extends Token
 	 */
 	public static int getAcModToken(PlayerCharacter pc, Equipment eq)
 	{
-		return eq.getACMod(pc).intValue();
+		return eq.getACMod(pc);
 	}
 
 	/**
@@ -271,7 +271,7 @@ public class EqContainerToken extends Token
 	 */
 	public static float getCarriedToken(Equipment eq)
 	{
-		return eq.numberCarried().floatValue();
+		return eq.numberCarried();
 	}
 
 	/**
@@ -314,7 +314,7 @@ public class EqContainerToken extends Token
 		{
 			return 0;
 		}
-		return eq.getContainedWeight(pc).floatValue();
+		return eq.getContainedWeight(pc);
 	}
 
 	/**
