+++ code/src/java/plugin/exporttokens/SkillSitToken.java
@@ -262,7 +262,7 @@ public class SkillSitToken extends Token
 
 				case SkillToken.SKILL_TOTAL:
 					int rank = SkillRankControl.getTotalRank(pc, skill).intValue()
-						+ SkillModifier.modifier(skill, pc).intValue();
+						+ SkillModifier.modifier(skill, pc);
 					if (isSituation)
 					{
 						rank += sit.getSituationBonus();
@@ -290,7 +290,7 @@ public class SkillSitToken extends Token
 					break;
 
 				case SkillToken.SKILL_MOD:
-					int mod = SkillModifier.modifier(skill, pc).intValue();
+					int mod = SkillModifier.modifier(skill, pc);
 					if (isSituation)
 					{
 						mod += sit.getSituationBonus();
@@ -307,7 +307,7 @@ public class SkillSitToken extends Token
 					break;
 
 				case SkillToken.SKILL_MISC:
-					int misc = SkillModifier.modifier(skill, pc).intValue();
+					int misc = SkillModifier.modifier(skill, pc);
 					if (isSituation)
 					{
 						misc += sit.getSituationBonus();
@@ -366,7 +366,7 @@ public class SkillSitToken extends Token
 					}
 					else
 					{
-						int mRank = etRank + SkillModifier.modifier(skill, pc).intValue();
+						int mRank = etRank + SkillModifier.modifier(skill, pc);
 						if (isSituation)
 						{
 							mRank += sit.getSituationBonus();
@@ -384,7 +384,7 @@ public class SkillSitToken extends Token
 					}
 					else
 					{
-						int mRank = tRank + SkillModifier.modifier(skill, pc).intValue();
+						int mRank = tRank + SkillModifier.modifier(skill, pc);
 						if (isSituation)
 						{
 							mRank += sit.getSituationBonus();
