+++ code/src/java/plugin/lsttokens/pcclass/MaxlevelToken.java
@@ -49,7 +49,7 @@ public class MaxlevelToken implements CDOMPrimaryToken<PCClass>
 			try
 			{
 				lim = Integer.valueOf(value);
-				if (lim.intValue() <= 0)
+				if (lim <= 0)
 				{
 					return new ParseResult.Fail("Value less than 1 is not valid for " + getTokenName() + ": " + value);
 				}
@@ -76,7 +76,7 @@ public class MaxlevelToken implements CDOMPrimaryToken<PCClass>
 		{
 			returnString = "NOLIMIT";
 		}
-		else if (lim.intValue() <= 0)
+		else if (lim <= 0)
 		{
 			context.addWriteMessage(getTokenName() + " must be an integer > 0");
 			return null;
