+++ code/src/java/plugin/lsttokens/spell/ClassesToken.java
@@ -141,11 +141,11 @@ public class ClassesToken extends AbstractTokenWithSeparator<Spell> implements C
 			try
 			{
 				level = Integer.valueOf(levelString);
-				if (level.intValue() < -1)
+				if (level < -1)
 				{
 					return new ParseResult.Fail(getTokenName() + " may not use a negative level: " + value);
 				}
-				else if (level.intValue() == -1)
+				else if (level == -1)
 				{
 					if (prereq != null)
 					{
@@ -302,7 +302,7 @@ public class ClassesToken extends AbstractTokenWithSeparator<Spell> implements C
 								"Incoming Allows Edge to " + spell.getKeyName() + " had no Spell Level defined");
 							return null;
 						}
-						if (level.intValue() < 0)
+						if (level < 0)
 						{
 							context.addWriteMessage("Incoming Allows Edge to " + spell.getKeyName()
 								+ " had invalid Level: " + level + ". Must be >= 0.");
